Move other-permissions into TableView
