+++ src/main/java/org/openbase/bco/bcozy/permissions/OtherPermissionsViewModel.java
@@ -1,49 +1,17 @@
 package org.openbase.bco.bcozy.permissions;
 
+import org.openbase.bco.bcozy.model.LanguageSelection;
+
 import java.util.logging.Logger;
 
 /**
  * @author vdasilva
  */
-public final class OtherPermissionsViewModel extends AbstractPermissionsViewModel {
+public final class OtherPermissionsViewModel extends AbstractTableRowPermissionsViewModel {
     private static final Logger LOG = Logger.getLogger(OtherPermissionsViewModel.class.getName());
 
-    private boolean read, write, access;
-
     public OtherPermissionsViewModel(boolean read, boolean write, boolean access) {
-        super(read, write, access);
-
-        this.read = (read);
-        this.write = (write);
-        this.access = (access);
+        super(LanguageSelection.getLocalized("other"), read, write, access);
     }
 
-    public boolean isRead() {
-        return read;
-    }
-
-    public void setRead(boolean read) {
-        this.read = read;
-    }
-
-    public boolean isWrite() {
-        return write;
-    }
-
-    public void setWrite(boolean write) {
-        this.write = write;
-    }
-
-    public boolean isAccess() {
-        return access;
-    }
-
-    public void setAccess(boolean access) {
-        this.access = access;
-    }
-
-    @Override
-    public boolean changed() {
-        return changed(isRead(), isWrite(), isAccess());
-    }
 }
